Migrate education page and components into new app directory - Update file naming to match /app conventions - FIX NEEDED: Refactor imports and file server vs client conventions
