+++ src/app/education/_components/Degree.tsx
@@ -1,4 +1,5 @@
 /** @format */
+"use client";
 
 import axios from "axios";
 import dynamic from "next/dynamic";
