+++ src/app/education/_components/DegreeCard.tsx
