+++ src/app/education/_components/EduBody.module.css
