+++ src/app/education/_components/EduImage.tsx
